refactor: move ts files to src/
