+++ src/algorithm.ts
