+++ src/binomial.ts
