+++ src/bit.ts
