+++ src/deque.ts
