+++ src/euclidean.ts
