+++ src/functional.ts
