+++ src/rolling-hash.ts
