+++ src/segment-tree.ts
