+++ src/trie.ts
